[#661] Add SOURCE_COMMIT env and build containers on Travis
The SOURCE_COMMIT contains the source's commit hash, to be used inside the
container. We had to move the container building to Travis to have more
flexibility on how the "docker build" command is run, as Docker Cloud's
automatic builds don't allow us to pass extra variables (only Docker Hub).
Another great side effefct is that the containers will only be built if the
tests pass on Travis.

opentrials/opentrials#661